refactor(cli): isolate mcp-dir server build
